backend/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.27
motor>=3.3
pymongo>=4.6
pydantic>=2.6,<3
pydantic-settings>=2.2
python-dotenv>=1.0
PyJWT>=2.8
passlib[bcrypt]>=1.7
bcrypt>=4.0,<5
python-multipart>=0.0.9
openpyxl>=3.1
rapidfuzz>=3.6
email-validator>=2.1
httpx>=0.27
anthropic>=0.39
